capsules/virtio: add VirtIO NetworkCard device driver
Implement a driver for the VirtIO NetworkCard (virtio-net)
device.

Given that no hardware interface layer for Ethernet network adapters
is defined in Tock as of this commit, it exposes a non-standard
interface. This device may, in conjuction with the LiteX LiteEth
device and potentially others help define an Ethernet HIL that is
compatible with a wide range of devices.

The feature VIRTIO_NET_F_MRG_RXBUF ("Driver can merge receive buffers"
[1]) is not negotiated, as it does not seem to be offered by QEMU
currently. Nonetheless, QEMU seems to be able to merge at least
separate header (12 byte) and packet packet buffers for receive
operations.

[1]: https://docs.oasis-open.org/virtio/virtio/v1.1/csprd01/virtio-v1.1-csprd01.html

Signed-off-by: Leon Schuermann <leon@is.currently.online>