qemu_rv32_virt: add VirtIO EntropySource & Net devices
Wire up the VirtIORng and VirtIONet drivers in the qemu_rv32_virt
board definition.

This adds an auto-discovery mechanism over the provided VirtIO MMIO
transport, which will detect supported VirtIO devices and allows to
dynamically instantiate the appropriate device drivers.

As no Ethernet HIL is defined in Tock as of this commit, VirtIO
NetworkCard devices are currently unused. A dummy driver to verify
basic device functionality is provided (commented out).

Signed-off-by: Leon Schuermann <leon@is.currently.online>